% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbc.R
\docType{package}
\name{adbc-package}
\alias{adbc}
\alias{adbc-package}
\title{adbc: Boilerplate for implementing a DBI back end}
\description{
A dummy DBI back end that provides access to a hypothetical data source named "Kazam", intended to be used as boilerplate for implementing DBI back ends from scratch.
}
\details{
TBD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkazam.r-dbi.org}
  \item \url{https://github.com/r-dbi/adbc}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
