% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_driver.R
\name{load_driver}
\alias{load_driver}
\title{Load a driver}
\usage{
load_driver(path, entrypoint = NULL)
}
\arguments{
\item{path}{Path to the shared library}

\item{entrypoint}{The name of the ADBC entry point,
by default \code{"AdbcDriverInit"} is used.}
}
\description{
TBD.
}
\examples{
try(load_driver("/path/to/foo.so"))
try(load_driver("/path/to/foo.so", "foo_adbc_init"))
}
