% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{adbc}
\alias{adbc}
\alias{from_radbc}
\title{adbc driver}
\usage{
adbc(path, entrypoint = NULL)

from_radbc(radbc)
}
\arguments{
\item{path}{Path to the shared library}

\item{entrypoint}{The name of the ADBC entry point,
by default \code{"AdbcDriverInit"} is used.}
}
\description{
TBD.
}
\examples{
\dontrun{
#' library(DBI)
adbc::adbc()
}
}
