% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_driver.R
\name{load_driver}
\alias{load_driver}
\title{Load a driver}
\usage{
load_driver(connection)
}
\arguments{
\item{connection}{A connection string of the form
\code{"Driver=<path-to-library>;Entrypoint=<entrypoint-name>"}}
}
\description{
TBD.
}
\examples{
try(load_driver("Driver=foo.so;Entrypoint=init_foo"))
}
